* Guarantee reading the expected number of bytes from the socket every time * Remove bufsize from client and conn, since they're not actually enforced
Notes:

This commit changes behavior a bit by raising a BufferUnderflowError when
no data is received for the message size rather than a ConnectionError.

Since bufsize in the socket is not actually enforced, but it is used by the consumer
when creating requests, moving it there until a better solution is implemented.