Check timeout type in KafkaClient constructor
If a future was passed as the only positional parameter it would
be assigned to the "timeout_ms" parameter erroneously. This mistake
would not raise any exception but would lead to odd behaviour later,
what could make it extremely difficult to troubleshoot.

Adding a type check ensures that an exception is raise earlier to
notify the user about the problem.